Move sounds to AlarmAudio
