+++ love/src/components/Layout/Layout.container.jsx
@@ -11,7 +11,6 @@ import {
   getLastComponentHeartbeat,
   getAllTime,
   getAllAlarms,
-  getLastestAlarms,
   getTaiToUtc,
 } from '../../redux/selectors';
 import { logout } from '../../redux/actions/auth';
@@ -34,7 +33,6 @@ const mapStateToProps = (state) => {
   const views = getViews(state);
   const timeData = getAllTime(state);
   const alarms = getAllAlarms(state);
-  const newAlarms = getLastestAlarms(state);
   const taiToUtc = getTaiToUtc(state);
   return {
     user,
@@ -47,7 +45,6 @@ const mapStateToProps = (state) => {
     views,
     timeData,
     alarms,
-    newAlarms,
     taiToUtc,
   };
 };
